feat: reorganized project
