+++ src/ffi/mod.rs
@@ -1,0 +1,2 @@
+mod boxed_slice;
+pub use boxed_slice::BoxedSlice;
